edit reference table

+++ database/migrations/2020_01_28_203720_create_tags_tables.php
@@ -16,7 +16,7 @@ class CreateTagsTables extends Migration
         Schema::create('tags', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('tagName')->nullable();
-            $table->integer('category_id')->unsigned()->index();
+            $table->bigInteger('category_id')->unsigned()->index();
             $table->integer('tag_count')->unsigned()->default(0);
         });
     }
